Update and rename myutils.js to utils.js
